requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[argon2]
python-jose[cryptography]
pydantic[email]
aiosqlite
